:hammer: extracted components for eventsView
