+++ features/shared/EventsView.css
@@ -1,30 +1,19 @@
-.highlightedEventsView {
+.eventsView {
+  width: 100vw;
+  max-width: 1450px; /* this is how wide all the category tabs are together */
   height: 100vh;
   display: flex;
   flex-direction: column;
 }
 
-.highlightedEventsView .p-tabview-panel {
+.eventsView .p-tabview-panel {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   gap: calc(var(--spacing) * 4);
 }
 
-.highlightedEventsView .eventsListsWrapper {
+.eventsView .eventsListsWrapper {
   padding: calc(var(--spacing) * 3) 0;
   flex-grow: 1;
   display: flex;
-}
-
-.highlightsGrid {
-  flex-grow: 1;
-  padding: calc(var(--spacing) * 2);
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: calc(var(--spacing) * 2);
-  align-content: start;
-}
-
-.highlightsGrid .eventCard {
-  max-height: 140px;
 }
