+++ pages/api/events.ts
@@ -19,90 +19,90 @@ export async function eventsEndpointHandler(
   request: ApiRequest,
 ): Promise<ApiResponse> {
   switch (request.method) {
-  case 'GET': {
-    try {
-      const events = await db.getAllAgendaEvents();
-      return { status: 200, message: 'OK', data: events };
-    } catch (error) {
-      return {
-        status: 500,
-        message: `Error on getAllAgendaEvents: ${error.message}`,
-      };
+    case 'GET': {
+      try {
+        const events = await db.getAllAgendaEvents();
+        return { status: 200, message: 'OK', data: events };
+      } catch (error) {
+        return {
+          status: 500,
+          message: `Error on getAllAgendaEvents: ${error.message}`,
+        };
+      }
     }
-  }
 
-  case 'POST': {
-    if (!validationService.newEvent(request.body)) {
+    case 'POST': {
+      if (!validationService.newEvent(request.body)) {
+        return {
+          status: 422,
+          message: 'Invalid new event object',
+          data: { providedObject: request.body },
+        };
+      }
+
+      try {
+        const dbResponse = await db.createNewAgendaEvent(request.body);
+        return { status: 200, message: 'OK', data: dbResponse };
+      } catch (error) {
+        return {
+          status: 500,
+          message: `Error on createNewAgendaEvent: ${error.message}.`,
+        };
+      }
+    }
+
+    case 'DELETE': {
+      if (!request?.body?.id) {
+        return {
+          status: 422,
+          message: 'Can not delete an event. Missing ID.',
+        };
+      }
+
+      try {
+        const dbResponse = await db.deleteAgendaEvent(request.body.id);
+        return { status: 200, message: 'OK', data: dbResponse };
+      } catch (error) {
+        return {
+          status: 500,
+          message: `Error on deleteAgendaEvent: ${error.message}.`,
+        };
+      }
+    }
+
+    case 'PUT': {
+      if (!request?.body?.id) {
+        return { status: 422, message: 'Can not update event. Missing ID.' };
+      }
+
+      if (Object.keys(request?.body).length < 2) {
+        return {
+          status: 422,
+          message: 'Can not update event. Missing values to be updated.',
+        };
+      }
+
+      try {
+        const dbResponse = await db.updateAgendaEvent(request.body);
+        return { status: 200, message: 'OK', data: dbResponse };
+      } catch (error) {
+        return {
+          status: 500,
+          message: `Error on updateAgendaEvent: ${error.message}.`,
+        };
+      }
+    }
+
+    default:
       return {
         status: 422,
-        message: 'Invalid new event object',
-        data: { providedObject: request.body },
+        message: 'Unsupported request method',
       };
-    }
-
-    try {
-      const dbResponse = await db.createNewAgendaEvent(request.body);
-      return { status: 200, message: 'OK', data: dbResponse };
-    } catch (error) {
-      return {
-        status: 500,
-        message: `Error on createNewAgendaEvent: ${error.message}.`,
-      };
-    }
-  }
-
-  case 'DELETE': {
-    if (!request?.body?.id) {
-      return {
-        status: 422,
-        message: 'Can not delete an event. Missing ID.',
-      };
-    }
-
-    try {
-      const dbResponse = await db.deleteAgendaEvent(request.body.id);
-      return { status: 200, message: 'OK', data: dbResponse };
-    } catch (error) {
-      return {
-        status: 500,
-        message: `Error on deleteAgendaEvent: ${error.message}.`,
-      };
-    }
-  }
-
-  case 'PUT': {
-    if (!request?.body?.id) {
-      return { status: 422, message: 'Can not update event. Missing ID.' };
-    }
-
-    if (Object.keys(request?.body).length < 2) {
-      return {
-        status: 422,
-        message: 'Can not update event. Missing values to be updated.',
-      };
-    }
-
-    try {
-      const dbResponse = await db.updateAgendaEvent(request.body);
-      return { status: 200, message: 'OK', data: dbResponse };
-    } catch (error) {
-      return {
-        status: 500,
-        message: `Error on updateAgendaEvent: ${error.message}.`,
-      };
-    }
-  }
-
-  default:
-    return {
-      status: 422,
-      message: 'Unsupported request method',
-    };
   }
 }
 
 // NextJS API lambda
-export default async function(
+export default async function (
   req: NextApiRequest,
   res: NextApiResponse,
 ): Promise<void> {
