+++ styles/components.css
@@ -1,3 +1,7 @@
+/* shared */
 @import '../features/shared/EventsLoadingProgressSpinner.css';
 @import '../features/shared/EventsLoadingError.css';
 @import '../features/shared/EventsView.css';
+
+/* highlighted events */
+@import '../features/view-highlighted-events/HighlightsList.css';
